Several enhancements including: Now you can see what set you are on after you focus is lost from the set list (treeview) . (Using a picture to indicate it). Now you can see what pictures are selected when a message box pops up. (The font is bold.)
I remove the top node to add extra space. Only the subfolders (subnodes) are added to the set list (treeview).

Now after an allTags scan completes, the set list will indicate which sets include photos with no DateTaken, or photos with a zero length tag. It does this by means of making the font bold for that set and assigning a tooltip. There are several things now that happen when the tagReader finishes and this is now being detected properly.