Enabled adding captions to photos.
Made it easier to add tags to more than one photo at a time.
Increments/decrements date by 1 second (depending on whether you navigate from the previous or the next photo) so that, if you click Set Date Taken for a photo with no date, you won't have two photos with the exact same date and time.

Now when selecting multiple photos, when you deselect a photo it shows a selected photo (usually the photo with the focus in the pictureList).

bug fix for .mov files