New feature allows user to sort their sets on flickr (a somewhat modified reverse alphabetical, to allow for dates). There's a new class called PhotosetComparer that supports this feature.
Now Videos tags (which reside on flickr only) are also being tracked in the allTags list. That means the background tag reader downloads flickr tags for everything now, so that process is slower. However, I think slowness is not a big issue because it is a background process.

More robust error handling on the local picture save.

Handles non UTF encoded tags a bit better now.

Now you can set the date taken for selected pictures. It will use the date of the FOCUSED picture and copy that to the other selected pictures. (The user will see visually the date/time being saved at confirmation.)

A couple of other minor bug fixes.